Increased 400% performance. Added load test.
PerformanceStats.java measures the performance of the threads. The method controls rage conditions through a synchronized method. This approach has been tested to a certain extend just for the purposes of this exercise (with and w/o the keyword with 20 processes, 10 thread pool threads and 2M requests), but would need further testing for production goals.